Move since-tag changes to the version update process
