+++ scripts/updateVersion.js
@@ -1,6 +1,7 @@
 const exec = require("child_process").exec;
 const path = require("path");
 const fs = require("fs");
+const glob = require("glob");
 const generateChangelog = require("./generateChangelog");
 
 const branchName = process.argv[2];
@@ -46,6 +47,31 @@ const updateEngineVersion = async (version) => {
     fs.writeFileSync(thinEngineFile, newThinEngineData);
 };
 
+const updateSinceTag = (version) => {
+    // get all typescript files in the dev folder
+    const files = glob.sync(path.join(baseDirectory, "packages", "dev", "**", "*.ts"));
+    files.forEach((file) => {
+        try {
+            // check if file contains @since\n
+            const data = fs.readFileSync(file, "utf-8").replace(/\r/gm, "");
+            if (data.indexOf("* @since\n") !== -1) {
+                console.log(`Updating @since tag in ${file} to ${version}`);
+                // replace @since with @since version
+                const newData = data.replace(
+                    /\* @since\n/gm,
+                    `* @since ${version}
+`
+                );
+
+                // write file
+                fs.writeFileSync(file, newData);
+            }
+        } catch (e) {
+            console.log(e);
+        }
+    });
+};
+
 async function runTagsUpdate() {
     await runCommand(
         `npx lerna version ${config.versionDefinition} --yes --no-push --conventional-prerelease --force-publish --no-private --no-git-tag-version ${
@@ -56,8 +82,12 @@ async function runTagsUpdate() {
     fs.rmSync("package-lock.json");
     await runCommand("npm install");
     const version = getNewVersion();
+    // update engine version
     await updateEngineVersion(version);
+    // generate changelog
     await generateChangelog(version);
+    // update since tags
+    updateSinceTag(version);
     if (dryRun) {
         console.log("skipping", `git commit -m "Version update ${version}"`);
         console.log("skipping", `git tag -a ${version} -m ${version}`);
